Camera movement finished. Add camera overlays

+++ Missions/DayZCommunityOfflineMode.ChernarusPlus/core/Module.c
@@ -70,7 +70,7 @@ class Module
 
 	// Override functions 
 	
-	void onUpdate( int timeslice ) 
+	void onUpdate( float timeslice ) 
 	{
 	}
